+++ Example/AuthExample/Assets.xcassets/AccentColor.colorset/Contents.json
