+++ Example/AuthExample/Assets.xcassets/AppIcon.appiconset/Contents.json
