+++ Example/AuthExample/Assets.xcassets/Contents.json
