+++ Example/AuthExample/ViewController.swift
@@ -1,0 +1,19 @@
+//
+//  ViewController.swift
+//  AuthExample
+//
+//  Created by Fumiya Tanaka on 2022/01/01.
+//
+
+import UIKit
+
+class ViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        // Do any additional setup after loading the view.
+    }
+
+
+}
+
