+++ Example/FirestoreExample/Info.plist
