+++ Example/FirestoreExample/ViewController+Async.swift
@@ -1,0 +1,14 @@
+//
+//  ViewController+Async.swift
+//  Example
+//
+//  Created by Fumiya Tanaka on 2022/01/01.
+//
+
+import Foundation
+
+extension ViewController {
+    func create_async() {
+
+    }
+}
